Create TaskPerformance
Instructions
Create a generic class called TriPrism. It shall store the dimensions of a triangular prism and calculate its volume. The type of the dimensions will be decided upon instantiation of the TriPrism objects. Use the formula: volume = length * (0.5 * base * height)

1.Create a generic class called TriPism. 
2.Declare three (3) private variables named length, base, and height.
3.Add three (3) pairs of getters and setters to retrieve and set the values of the variables mentioned above.
4.Create another class named TriPrismDemo.
5.In its main method, instantiate two (2) TriPrism objects named tp1 (Integer) and tp2 (Double).
6.Use the setter methods to set the values of the dimensions. 
Calculate and display the volume of the two (2) TriPrism objects using the getter methods. Refer to the formula provided above.